Desktop/kazakhstanBotUpdate/KazakhstanBot/kazakhstanBot-master/requirements.txt
discord.py
asyncpg
aiohttp
python-dotenv
pytz
